scripts/texture_generator/requirements.txt
click
colorama
pillow
